Rename config file, add guide in README
The config file shouldn't be named simply "config.yml" when it lives in
the system configuration place. Or, if it is, it should live in a sub-
folder named to indicate the owning application.

I've renamed the expected file to "/etc/unifi-monitor.yml", and added
some usage guidance to the README.md.